+++ CHANGELOG.md
@@ -18,6 +18,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - "source" property emitted in events. "provider" has to be used instead.
 - "sources" handler. "instances" has to be used instead.
 
+## [1.8.0] - 2020-01-18
+### Added
+- Add `cleanState` method
+
+### Changed
+- Update dependency husky to v4.0.10
+- Update dependency babel monorepo to v7.8.3
+
 ## [1.7.3] - 2020-01-13
 ### Changed
 - Update dependencies